finished project

+++ backend/src/main/java/edu/bbardi/pokerbackend/security/dto/JwtResponse.java
@@ -12,6 +12,6 @@ public class JwtResponse {
     private Long id;
     private String username;
     private String email;
-    private Float balance;
+    private Long balance;
     private List<String> roles;
 }
